Move new components/functions into new file
